Fix Resources folder not getting build bug
