+++ src/main/java/com/oodwj_assignment/Controllers/cusHistoryController.java
@@ -1,0 +1,4 @@
+package com.oodwj_assignment.Controllers;
+
+public class cusHistoryController {
+}
