+++ src/main/java/com/oodwj_assignment/Controllers/cusNotificationController.java
@@ -1,4 +1,4 @@
-package com.oodwj_assignment;
+package com.oodwj_assignment.Controllers;
 
 public class cusNotificationController {
 }
