+++ src/main/java/com/oodwj_assignment/Controllers/cusProfileController.java
@@ -1,0 +1,4 @@
+package com.oodwj_assignment.Controllers;
+
+public class cusProfileController {
+}
