+++ src/main/java/com/oodwj_assignment/Controllers/cusWalletController.java
@@ -1,0 +1,4 @@
+package com.oodwj_assignment.Controllers;
+
+public class cusWalletController {
+}
