--- src/main/java/com/oodwj_assignment/cusProfileController.java
@@ -1,4 +1,0 @@
-package com.oodwj_assignment;
-
-public class cusProfileController {
-}
