--- src/main/java/com/oodwj_assignment/cusReviewController.java
@@ -1,4 +1,0 @@
-package com.oodwj_assignment;
-
-public class cusReviewController {
-}
