+++ src/main/java/module-info.java
@@ -5,5 +5,7 @@ module com.oodwj_assigment {
 
     opens com.oodwj_assignment to javafx.fxml;
     exports com.oodwj_assignment;
+    exports com.oodwj_assignment.Controllers;
+    opens com.oodwj_assignment.Controllers to javafx.fxml;
 
 }
